refactor: review project structure
