+++ cpu.go
@@ -1,0 +1,7 @@
+package systatus
+
+import "net/http"
+
+type CPUResponse struct{}
+
+func HandleCPU(w http.ResponseWriter, r *http.Request) {}
