+++ disk.go
@@ -1,0 +1,5 @@
+package systatus
+
+import "net/http"
+
+func HandleDisk(w http.ResponseWriter, r *http.Request) {}
